+++ TabletopTweaks/NewComponents/Prerequisites/PrerequisiteSpellKnown.cs
@@ -1,26 +1,31 @@
 ﻿using JetBrains.Annotations;
+using Kingmaker.Blueprints;
 using Kingmaker.Blueprints.Classes.Prerequisites;
 using Kingmaker.Blueprints.JsonSystem;
 using Kingmaker.UnitLogic;
+using Kingmaker.UnitLogic.Abilities.Blueprints;
 using Kingmaker.UnitLogic.Class.LevelUp;
 using System.Linq;
 
 namespace TabletopTweaks.NewComponents.Prerequisites {
-    [TypeId("08d2e61c79c64ee1afdca9fc834ffc32")]
-    class PrerequisiteMythicSpellbook : Prerequisite {
+    [TypeId("e523c74b8da74fec91ae651138ec0ca0")]
+    class PrerequisiteSpellKnown : Prerequisite {
+        private BlueprintAbility Spell {
+            get {
+                if (m_Spell == null) {
+                    return null;
+                }
+                return m_Spell.Get();
+            }
+        }
         public override bool CheckInternal([CanBeNull] FeatureSelectionState selectionState, [NotNull] UnitDescriptor unit, [CanBeNull] LevelUpState state) {
             return unit.Spellbooks
-                .Where(book => book.IsMythic)
-                .Any(book => book.MaxSpellLevel >= RequiredSpellLevel);
+                .Any(book => book.IsKnown(Spell));
         }
 
         public override string GetUITextInternal(UnitDescriptor unit) {
-            if (RequiredSpellLevel > 0) {
-                return $"Can cast spells of level {RequiredSpellLevel} or higher from mythic spellbook";
-            }
-            return $"Has mythic spellbook";
+            return $"Can cast spell: {Spell.Name}";
         }
-
-        public int RequiredSpellLevel;
+        public BlueprintAbilityReference m_Spell;
     }
 }