+++ TabletopTweaks/NewComponents/Prerequisites/PrerequisiteSpellbook.cs
@@ -1,11 +1,13 @@
 ﻿using JetBrains.Annotations;
 using Kingmaker.Blueprints;
 using Kingmaker.Blueprints.Classes.Prerequisites;
+using Kingmaker.Blueprints.JsonSystem;
 using Kingmaker.UnitLogic;
 using Kingmaker.UnitLogic.Class.LevelUp;
 using System.Linq;
 
 namespace TabletopTweaks.NewComponents.Prerequisites {
+    [TypeId("7686e2d0ab864daaaf01150c62741aba")]
     class PrerequisiteSpellbook : Prerequisite {
         public override bool CheckInternal([CanBeNull] FeatureSelectionState selectionState, [NotNull] UnitDescriptor unit, [CanBeNull] LevelUpState state) {
             return unit.Spellbooks