+++ TabletopTweaks/NewComponents/Prerequisites/PrerequisiteStatValues.cs
@@ -1,4 +1,5 @@
 ﻿using Kingmaker.Blueprints.Classes.Prerequisites;
+using Kingmaker.Blueprints.JsonSystem;
 using Kingmaker.Blueprints.Root;
 using Kingmaker.Designers.Mechanics.Facts;
 using Kingmaker.EntitySystem.Stats;
@@ -8,6 +9,7 @@ using System.Linq;
 using System.Text;
 
 namespace TabletopTweaks.NewComponents.Prerequisites {
+    [TypeId("8dccf39766ed482eb9b7f2ae31e50428")]
     class PrerequisiteStatValues : Prerequisite {
         public override bool CheckInternal(FeatureSelectionState selectionState, UnitDescriptor unit, LevelUpState state) {
             return CheckUnit(unit);